Split resources from API
